requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-bootstrap-v5==1.0.5
django-stdimage==5.3.0
gunicorn==20.1.0
importlib-metadata==3.10.1
Pillow==8.3.1
PyMySQL==1.0.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.3.0
zipp==3.5.0
